other-scripts/requirements.txt
requests
python-dotenv
boto3
opensearch-py
requests-aws4auth
